Add error_handlers
